Polymorphic working yey
Adding polymorphic_list=True in a base polymorphic admin model allows to properly get child classes in the admin interface.
I added it to the relevant classes and used __str__ to show nice names on the web.
This is technically very unoptimized with regards to DB usage, but since it's used for things that probably won't pass the tens of objects all together, I doubt it matters. Citation needed :)